pyproject.toml
[tool.poetry]
name = "no-pain"
version = "0.1.0"
description = "Doctor working hours"
authors = ["karol <karol.misiarz@robovision.eu>"]
readme = "README.md"
packages = [{include = "backend"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.95.1"
uvicorn = "0.22.0"
jinja2 = "3.1.2"
aiofile = "3.8.5"
python-dotenv = "1.0.0"
sqlalchemy = "2.0.13"
passlib = "1.7.4"
email-validator = "2.0.0.post2"
python-jose = "3.3.0"
python-multipart = "0.0.6"
fastapi-sessions = "^0.3.2"


[tool.poetry.group.dev.dependencies]
black = "23.3.0"
mypy = "1.3.0"
isort = "5.12.0"
pre-commit = "3.3.1"


[tool.poetry.group.test.dependencies]
pytest = "7.3.1"
requests = "2.30.0"
httpx = "0.24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
